refactor(refactors_and_fixes): change cache param from created to updated_at
